TLUT flow support for verilog using Yosys

+++ python/src/mapping.py
@@ -394,8 +394,16 @@ def synthesize(top, K, verboseFlag=False, workFiles=[]):
 
     workFiles = ' '.join(workFiles)
     print(workFiles)
-    cmd =  "docker run --rm -t   -v $(pwd):/src   -w /src   ghdl/synth:beta   yosys -m ghdl -p 'ghdl -fsynopsys "+workFiles+ " -e "+basename+"; synth -lut "+str(K)+" ; write_blif "+basename+".blif'"
-
+    if ext == "vhd" or ext == "vhdl":
+        cmd =  "docker run --rm -t   -v $(pwd):/src   -w /src   ghdl/synth:beta   yosys -m ghdl -p 'ghdl -fsynopsys "+workFiles+ " -e "+basename+"; synth -lut "+str(K)+" ; write_blif "+basename+".blif'"
+    elif ext == "v":
+        with open("synthesis.ys", "w") as sfile:
+            sfile.write("read_verilog "+workFiles+"\n")
+            sfile.write("synth -lut "+str(K)+"\n")
+            sfile.write("write_blif "+basename+".blif")         
+        cmd = "yosys synthesis.ys"
+    else:
+        raise Exception ('Yosys: HDL type not supported')
     if verboseFlag:
         print cmd
     output = commands.getoutput(cmd);
